Remove constraint on CPU realization index
In a6bc41e6ba2080d70805a370e62f9dfc6868ad20 we added a limitation on how
threads get realized because the `threadcount` test was failing.  It failed
because we attempted to access a DMA element outside of the allocated range.
However, we also constrained the DMA behavior to sysemu for the time being.

This check in cpu_realize() causes failures in linux-userspace test scenarios
like calling pthread_create().  This is undesirable because linux userspace
programs should be able to create many software threads, not bound by the
total number of schedulable vCPUs.  Linux userspace test cases unnaturally
fail with this error case.

Signed-off-by: Brian Cain <bcain@quicinc.com>